refactor(sdks/go): rename chunk_test.go→batch_test.go; fold value_kind_test.go into value_test.go (refs #213)
