+++ sdks/go/batch_test.go
